sql: user defined function schema changer elements definition
This commit adds the element protobuf definitions for functions
so that declarative schema changer would be able to decompose
or create a function descriptor.
The philosophy is that anything of function that is alterable
or can be altered through a `create or replace function` statement
is defined as a element.

Release note: None.